Add e2e test for manual prior notification form freeze during send
