+++ frontend/cypress/e2e/side_window/manual_prior_notification_form/utils.ts
@@ -1,44 +1,9 @@
-import { SideWindowMenuLabel } from 'domain/entities/sideWindow/constants'
+import { openSideWindowPriorNotificationListAsSuperUser } from '../prior_notification_list/utils'
 
-export const addSideWindowPriorNotification = () => {
-  cy.viewport(1920, 1080)
-  cy.visit('/side_window')
-  cy.wait(500)
-  if (document.querySelector('[data-cy="first-loader"]')) {
-    cy.getDataCy('first-loader').should('not.be.visible')
-  }
-
-  cy.clickButton(SideWindowMenuLabel.PRIOR_NOTIFICATION_LIST)
-  if (document.querySelector('[data-cy="first-loader"]')) {
-    cy.getDataCy('first-loader').should('not.be.visible')
-  }
-  cy.get('.Table-SimpleTable tr').should('have.length.to.be.greaterThan', 0)
+export const addManualSideWindowPriorNotification = () => {
+  openSideWindowPriorNotificationListAsSuperUser()
 
   cy.clickButton('Ajouter un préavis')
-  if (document.querySelector('[data-cy="first-loader"]')) {
-    cy.getDataCy('first-loader').should('not.be.visible')
-  }
-}
-
-export const editSideWindowPriorNotification = (vesselName: string, reportId: string) => {
-  cy.viewport(1920, 1080)
-  cy.visit('/side_window')
-  cy.wait(500)
-  if (document.querySelector('[data-cy="first-loader"]')) {
-    cy.getDataCy('first-loader').should('not.be.visible')
-  }
-
-  cy.clickButton(SideWindowMenuLabel.PRIOR_NOTIFICATION_LIST)
-  if (document.querySelector('[data-cy="first-loader"]')) {
-    cy.getDataCy('first-loader').should('not.be.visible')
-  }
-  cy.get('.Table-SimpleTable tr').should('have.length.to.be.greaterThan', 0)
-
-  cy.get('[data-cy="side-window-sub-menu-ALL"]').click()
-  cy.fill('Rechercher un navire', vesselName)
-
-  // TODO Allow for `string` type in monitor-ui.
-  cy.getTableRowById(reportId as any).clickButton('Éditer le préavis')
   if (document.querySelector('[data-cy="first-loader"]')) {
     cy.getDataCy('first-loader').should('not.be.visible')
   }
