+++ frontend/cypress/e2e/side_window/prior_notification_list/utils.ts
@@ -1,16 +1,15 @@
 import { SideWindowMenuLabel } from '../../../../src/domain/entities/sideWindow/constants'
+import { openSideWindowAsSuperUser, openSideWindowAsUser } from '../utils'
 
-export const openSideWindowPriorNotificationList = () => {
-  cy.viewport(1920, 1080)
-  cy.visit('/side_window')
-  cy.wait(500)
-  if (document.querySelector('[data-cy="first-loader"]')) {
-    cy.getDataCy('first-loader').should('not.be.visible')
-  }
+export const openSideWindowPriorNotificationListAsUser = openSideWindowAsUser
+
+export const openSideWindowPriorNotificationListAsSuperUser = () => {
+  openSideWindowAsSuperUser()
 
   cy.clickButton(SideWindowMenuLabel.PRIOR_NOTIFICATION_LIST)
   if (document.querySelector('[data-cy="first-loader"]')) {
     cy.getDataCy('first-loader').should('not.be.visible')
   }
+
   cy.wait(1000)
 }
