+++ frontend/package.json
@@ -37,7 +37,7 @@
   "dependencies": {
     "@dnd-kit/core": "6.1.0",
     "@dnd-kit/modifiers": "6.0.1",
-    "@mtes-mct/monitor-ui": "21.0.0",
+    "@mtes-mct/monitor-ui": "21.0.1",
     "@reduxjs/toolkit": "2.2.7",
     "@sentry/react": "7.117.0",
     "@tanstack/react-table": "8.19.3",
